Bug Fixed
修复金蝶模式九宫格嬗变不升级：升级次数大于18时会出现嬗变但不升级的情况
修复金蝶模式NPC“夏渔雾溟”的次数给予是重置当前次数为2，而非添加2次
修复金蝶模式20层Roll到矮人将军楼层上下楼失效
修复萨卡班甲鱼楼层贴图渲染异常
修复重型弩炮伤害类型异常
修复创世神之心在特定情况下的活死人严重异常
金蝶模式三饰品将在首次对话“夏渔雾溟”直接给予

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/Level.java
@@ -342,13 +342,9 @@ public abstract class Level implements Bundlable {
 				addItemToSpawn( new Stylus() );
 			}
 
-			if ( Dungeon.trinketCataNeeded() ){
+			if ( Dungeon.trinketCataNeeded() && !Statistics.RandMode ){
 				Dungeon.LimitedDrops.TRINKET_CATA.drop();
 				addItemToSpawn( new TrinketCatalyst());
-				if(Statistics.RandMode){
-					addItemToSpawn( new TrinketCatalyst());
-					addItemToSpawn( new TrinketCatalyst());
-				}
 			}
 			//one scroll of transmutation is guaranteed to spawn somewhere on chapter 2-4
 			int enchChapter = (int)((Dungeon.seed / 10) % 3) + 1;
